requirements.txt
asgiref==3.8.1
attrs==25.3.0
beeprint==2.4.11
certifi==2025.1.31
csscompressor==0.9.5
Django==5.2
elasticsearch==6.3.1
fusepy==3.0.1
gitignore_parser==0.1.12
h11==0.14.0
idna==3.10
outcome==1.3.0.post0
psutil==4.3.1
pyScss==1.4.0
PySocks==1.7.1
selenium==4.31.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.3
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.13.2
urllib3==2.4.0
urwid==2.6.16
wcwidth==0.2.13
websocket-client==1.8.0
wheel==0.32.3
wsproto==1.2.0
